requirements.txt
numpy
matplotlib
ipykernel
scipy
PyYAML
# only for KDE
PyQt5
PyQt6